+++ src/index.js
@@ -1,21 +1,19 @@
 #!/usr/bin/env node
 
 const yargs = require('yargs')
+const { noop } = require('lodash')
 
-const { mainProcess } = require('./cli')
+const mainProcess = require('./cli')
 
 yargs
   .usage('Usage: $0 <command> [options]')
-  .command(
-    '$0 [template]',
-    'Generate README.md from a template',
-    command =>
-      command.positional('template', {
-        desc: 'The name of template you want to use',
-        default: 'default'
-      }),
-    args => mainProcess(args)
-  )
+  .command('$0', 'Generate README.md', noop, args => {
+    const { path: customTemplatePath, yes: useDefaultAnswers } = args
+    mainProcess({ customTemplatePath, useDefaultAnswers })
+  })
+  .string('p')
+  .alias('p', 'path')
+  .describe('path', 'Path to your own template')
   .boolean('yes')
   .alias('y', 'yes')
   .describe('yes', 'Use default values for all fields')