+++ src/questions/license-url.spec.js
@@ -5,26 +5,42 @@ describe('askLicenseUrl', () => {
     const licenseUrl =
       'https://github.com/kefranabg/readme-md-generator/blob/master/LICENSE'
     const projectInfos = { licenseUrl }
-    const answersContext = { licenseName: 'MIT' }
 
-    const result = askLicenseUrl(projectInfos, answersContext)
+    const result = askLicenseUrl(projectInfos)
 
-    expect(result).toEqual({
-      type: 'input',
-      message: '📝  License url (use empty value to skip)',
-      name: 'licenseUrl',
-      default: licenseUrl
-    })
+    expect(result).toEqual(
+      expect.objectContaining({
+        type: 'input',
+        message: '📝  License url (use empty value to skip)',
+        name: 'licenseUrl',
+        default: licenseUrl
+      })
+    )
   })
 
-  it('should return undefined', () => {
-    const licenseUrl =
-      'https://github.com/kefranabg/readme-md-generator/blob/master/LICENSE'
-    const projectInfos = { licenseUrl }
-    const answersContext = { licenseName: '' }
+  it('should show this question if licenseName is defined', () => {
+    const projectInfos = {
+      licenseUrl:
+        'https://github.com/kefranabg/readme-md-generator/blob/master/LICENSE'
+    }
+    const answersContext = { licenseName: 'MIT' }
 
-    const result = askLicenseUrl(projectInfos, answersContext)
+    const question = askLicenseUrl(projectInfos)
+    const result = question.when(answersContext)
 
-    expect(result).toBe(undefined)
+    expect(result).toBe(true)
+  })
+
+  it('should not show this question if licenseName is not defined', () => {
+    const projectInfos = {
+      licenseUrl:
+        'https://github.com/kefranabg/readme-md-generator/blob/master/LICENSE'
+    }
+    const answersContext = {}
+
+    const question = askLicenseUrl(projectInfos)
+    const result = question.when(answersContext)
+
+    expect(result).toBe(false)
   })
 })